+++ Circles/Views/Groups/GroupSettingsView.swift
@@ -152,6 +152,10 @@ struct GroupSettingsView: View {
                 }
             }
             
+            if debugMode {
+                RoomDebugDetailsSection(room: room)
+            }
+            
             Section("Danger Zone") {
                 Button(role: .destructive, action: {
                     self.showConfirmLeave = true